Moved stuff for orginization
