+++ src/main/java/edu/xavier/csci/flowerpower/Reservation/ReservationController.java
@@ -1,10 +1,8 @@
-package edu.xavier.csci.flowerpower;
+package edu.xavier.csci.flowerpower.Reservation;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
-import java.util.*;
 
 @RestController
 @CrossOrigin
