+++ src/main/java/edu/xavier/csci/flowerpower/Reservation/ReservationRepository.java
@@ -1,4 +1,4 @@
-package edu.xavier.csci.flowerpower;
+package edu.xavier.csci.flowerpower.Reservation;
 
 import org.springframework.data.repository.CrudRepository;
