+++ src/main/java/edu/xavier/csci/flowerpower/Reservation/ReservationService.java
@@ -1,4 +1,4 @@
-package edu.xavier.csci.flowerpower;
+package edu.xavier.csci.flowerpower.Reservation;
 
 import org.springframework.stereotype.Service;
 import java.util.ArrayList;
